Extract Program list to another component
